# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"My API Docs","description":"My API","termsOfService":"http:\/\/localhost.io\/terms\/","contact":{"email":"apiteam@localhost.io"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"paths":{"\/api\/auth\/login":{"post":{"tags":["auth"],"summary":"Login","description":"Gets user API token.","operationId":"login","requestBody":{"description":"Returns API token with given user email and password","required":true,"content":{"application\/json":{"schema":{"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"properties":{"authorization":{"type":"string"}},"type":"object"}}}},"401":{"description":"Access denied"},"403":{"description":"Inactive account"},"404":{"description":"Not found"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}}}},"\/api\/auth\/forgot-password":{"post":{"tags":["auth"],"summary":"Forgot password","description":"Sends an email to reset user password.","operationId":"forgotPassword","requestBody":{"description":"user email","required":true,"content":{"application\/json":{"schema":{"properties":{"email":{"type":"string","format":"email"}},"type":"object"}}}},"responses":{"200":{"description":"successful operation"},"403":{"description":"Inactive account"},"404":{"description":"Not found"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}}}},"\/api\/auth\/reset-password":{"get":{"tags":["auth"],"summary":"Reset password","description":"Resets user password and sends new password to user.","operationId":"resetPassword","parameters":[{"name":"key","in":"query","description":"Unique key to reset password","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","description":"email of user","required":true,"schema":{"type":"string","format":"email"}}],"responses":{"200":{"description":"successful operation"},"401":{"description":"Access denied"},"403":{"description":"Inactive account"},"404":{"description":"Not found"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}}}},"\/api\/auth\/register":{"post":{"tags":["auth"],"summary":"Register user","description":"Returns registered user.","operationId":"register","requestBody":{"description":"User details will be used to create. Email is mandatory.","required":true,"content":{"application\/json":{"schema":{"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"}},"type":"object"}}}},"responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"412":{"description":"Email exists"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}}}},"\/api\/auth\/activate":{"get":{"tags":["auth"],"summary":"Activate user","description":"Activates user with given key and email and send welcome email.","operationId":"activate","parameters":[{"name":"key","in":"query","description":"Unique activation key","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","description":"email of user","required":true,"schema":{"type":"string","format":"email"}}],"responses":{"200":{"description":"successful operation"},"401":{"description":"Access denied"},"403":{"description":"Inactive account"},"404":{"description":"Not found"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}}}},"\/api\/auth\/resend-activation":{"post":{"tags":["auth"],"summary":"Resend activation","description":"Re-sends an email to activate user.","operationId":"resendActivation","requestBody":{"description":"user email","required":true,"content":{"application\/json":{"schema":{"properties":{"email":{"type":"string","format":"email"}},"type":"object"}}}},"responses":{"200":{"description":"successful operation"},"401":{"description":"Access denied"},"403":{"description":"Inactive account"},"404":{"description":"Not found"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}}}},"\/api\/user":{"get":{"tags":["user"],"summary":"Get all users","description":"Returns users list. Admin only","operationId":"get","responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"type":"array","items":{"$ref":"#\/components\/schemas\/UserModel"}}}}},"403":{"description":"Unauthorized"},"500":{"description":"Server error"}},"security":[{"Authorization":[]}]},"post":{"tags":["user"],"summary":"Create user","description":"Returns created user. Admin only","operationId":"post","requestBody":{"description":"Create user object","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"403":{"description":"Unauthorized"},"412":{"description":"Email exists"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}},"security":[{"Authorization":[]}]}},"\/api\/user\/{id}":{"get":{"tags":["user"],"summary":"Get the user","description":"Returns single user with given ID. Admin only","operationId":"show","parameters":[{"name":"id","in":"path","description":"ID of user to return","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"403":{"description":"Unauthorized"},"404":{"description":"Not found"},"500":{"description":"Server error"}},"security":[{"Authorization":[]}]},"put":{"tags":["user"],"summary":"Update user","description":"Returns updated user. Admin only","operationId":"put","parameters":[{"name":"id","in":"path","description":"ID of user","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"Update user object","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"403":{"description":"Unauthorized"},"404":{"description":"Not found"},"422":{"description":"Invalid input"},"500":{"description":"Server error"}},"security":[{"Authorization":[]}]},"delete":{"tags":["user"],"summary":"Delete user","description":"Admin only","operationId":"delete","parameters":[{"name":"id","in":"path","description":"ID of user","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation"},"403":{"description":"Unauthorized"},"404":{"description":"User not found"},"500":{"description":"Server error"}},"security":[{"Authorization":[]}]}},"\/api\/user\/me":{"get":{"tags":["user"],"summary":"Get current user","description":"Returns authenticated user.","operationId":"me","responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/UserModel"}}}},"403":{"description":"Unauthorized"},"500":{"description":"Server error"}},"security":[{"Authorization":[]}]}}},"components":{"schemas":{"UserGender":{"type":"string","enum":["male","female"]},"UserRole":{"type":"string","default":"user","enum":["user","admin"]},"UserStatus":{"type":"boolean","default":"false"},"UserMetaModel":{"properties":{"id":{"type":"integer"},"user_id":{"type":"integer"},"gender":{"type":"string","$ref":"#\/components\/schemas\/UserGender"},"phone":{"type":"string"},"city":{"type":"string"},"address":{"type":"string"},"birth_date":{"type":"string","format":"datetime"},"created_at":{"type":"string","format":"datetime"},"updated_at":{"type":"string","format":"datetime"}},"type":"object"},"UserModel":{"properties":{"id":{"type":"integer"},"first_name":{"type":"string"},"last_name":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"},"role":{"type":"string","$ref":"#\/components\/schemas\/UserRole"},"api_token":{"type":"string"},"status":{"type":"boolean","$ref":"#\/components\/schemas\/UserStatus"},"created_at":{"type":"string","format":"datetime"},"updated_at":{"type":"string","format":"datetime"}},"type":"object"}},"securitySchemes":{"Authorization":{"type":"apiKey","description":"User token to be used to access API end points","name":"Authorization","in":"header"}}},"externalDocs":{"description":"Find out more about My API","url":"http:\/\/localhost.io"}}